requirements.txt
flask
ffmpeg-python
opencv-python
numpy
Pillow
torch
transformers
flasgger
torch
transformers
Pillow
torchaudio 
datasets
librosa
